Added 1 test, reformatted a bit the code for better modularization, removed unused imports
